pyproject.toml
[project]
name = "cool_steg"
version = "0.1.0"
description = "Hide your messages or files inside an image and send it to your crush"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "python-magic-bin>=0.4.14",
    "pyzbar>=0.1.9",
    "qrcode>=8.2",
    "rich>=14.3.3",
    "typer>=0.24.1",
]
[project.scripts]
cool_steg = "cool_steg.cli:main"
